+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 ## [Unreleased]
 
 - (`16462ae`) Report top-level for-of statements.
+- (`dc5f99e`) Improve performance of `no-top-level-side-effect`.
 
 ## [0.2.2] - 2022-12-30