+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 
 - (`e97db34`) Report top-level for-in statements.
 - (`16462ae`) Report top-level for-of statements.
+- (`dc5f99e`) Improve performance of `no-top-level-side-effect`.
 
 ## [0.2.2] - 2022-12-30