+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning].
 - (`e97db34`) Report top-level for-in statements.
 - (`16462ae`) Report top-level for-of statements.
 - (`0a3ba0f`) Report top-level throw statements.
+- (`0fb75fa`) Report top-level try statements.
 - (`dc5f99e`) Improve performance of `no-top-level-side-effect`.
 
 ## [0.2.2] - 2022-12-30